First pass for new mapping registry
This registry provides means for registering mappings from one type to
another. Some existing mapping libraries use reflection to figure out
what fields exist and where to map them to, however this cannot be
applied here sadly.

For one, the types the mapping should be based on - the interfaces - do
not provide any means of figuring out which fields exist. This would
work at runtime, since there are properties to get the representation
of an object, but that is too late for us. Secondly, We'd need to
specify multiple things anywany, such as the creation of the objects, as
they come from factories, as well as how to map different types of
elements inside a list to different properties on the mapped element.

Therefor, the registry is a little more static, which means it needs to
be defined prior how properties from object A get mapped to object B.
This looks like this:

```xtend
registry.configure(sourceType, targetType) [
    create = [ Factory.eINSTANCE.createTargetType ]
    when = [ <sourceType predicate> ]
    map(getter).thenSet [ targetElement, obj |
        <update target element with obj>
    ]
    mapAll(listGetter).thenSet [ targetElement, list |
        <update target element with list>
    ]
]
```

The `map()` and `mapAll()` calls can be repeated as often as needed and
can be passed a filtered list of the original.

One might wonder why we need a predicate, as it should be a simple A->B
mapping. Sadly, it's not. There are some cases where in the source model
we have one type while in the target model we have different types to
represent something. Thus when mapping, we need to decide which element
in the target we need to use and that's the job of the predicate to
decide. This is the case, for example, the case with composed data types
as either the `type` is set or the `reference` is set.

There's one part that is missing here: how to resolve references. In the
textual language, you can reference other elements by name and it will
make sure they exist. When this textual model gets parsed into the model
in memory, these references no longer exist as just a reference, they'll
be replaced with a reference to the actual object in memory. Thus we
don't actually know when we're just dealing with a reference or the
definition of the object without knowing the technical background of an
element, i.e. is it possible for an element to be defined at a specific
location or does that location always imply it's a reference.

We need to know this so we can properly assign the element to the target
model as that needs to be aware of references in order to determine if
it should write out the element or write out a reference when it's
saving to disk. Currently, this is not completely solved as that
information is not easily attainable all the time. In most cases, this
information comes from the technical knowledge of the element where, at
the time of registration, we know if we are dealing with references or
we're dealing with new definitions. Thankfully so far, in those cases
where it would matter, it turned out that it actually didn't matter as
only the overall context mattered, which was the same. Specifically,
this can be observed with the DataType(s), as it's not clear if the
datatype used inside a composite datatype was defined at that place or
if it was defined outside, inside the scope of the repository.

References are thus left over to the registration to determine. Though
this brough up a different problem. Given that references to objects are
represented as actual references in memory, we need to make sure that is
the same for the mapped objects. Unfortunately, not all elements provide
us with an identification - a Fragement has a name, however the Content
subtypes all define their own identifier. But, since we're dealing with
the same exact object, we can just use the instance IDs of objects that
java assigns to them, because it would always return the same number.