Split out implementations and provde after callback
This splits out the implementations for multiple interfaces into their
own code files, instead of residing next to the interface in the same
file. This makes it easier to find what you're looking for.

Additionally, this adds a callback to the registration that can be
configured. This callback will be run after all the mappings have been
applied to allow for updating references between elements or similar.