Cache mapped objects by id and target type
Previously we only cached mapped objects by their ID. This however does
not work well if we - possibly - have the same object mapped as two
different target types. While this should not happen, this makes sure
that we still deal with this case properly.

Now instead of retuning the previous mapped object, if the target type
is specified differently, it will not find the cached object and will
instead run the mapping normally again.