Split configuration of generator into new bundle
This moves the configuration of the generator into a new bundle,
currently called `configuration`. To facilitate this, the main bundle
now provides an extension point `configure` with an interface called
`TransformationRegistryConfigurer`. This extension point can then be
used to setup the registry by configuring the individual mappings in the
provided registry.