Do not use xtext importer because we don't need it
Initially, it was implemented to support importing other modelling
files, to facilitate the usage of predefined types. However, this is not
required as instead we will let EMF do the resolving internally. This
was mostly done because it was expected that the scope needs to be
extended to support this. However, since already existing elements will
be used instead, the scope does not need to be increased. As a result,
this can be made much simpler.