Allow for specifying existing objects in mappings
This allows for adding objects from other models into a text model,
such as objects from other files. This is especially important since
there are specific objects that are used internally to do calculations.
 Currently, this resolution is quite barebones; you specify the same
absolute path as emf would use inside a resource file. This could in
the future be extended to allow for shortcuts to base palladio types.
However, to keep the grammar clean and avoid adding special cases to
the generator, this is not yet implemented.