Add more validation for registration and runtime
This adds two more places for validating the transformation. One being
at time of registration to check whether another "wildcard"
transformation has already been configured, i.e. a transformation that
doesn't have a `when` directive. If two are registered with the same
source and target type, we'd not know which one to use. We include the
target here too because there are cases where we specifically want
certain transformations. This needs to be possible and therefor we need
to include the target, however, at runtime one has to make sure to tell
the transformer which target type one wants. Additionally, it makes sure
that a factory has been configured, as we cannot operate without one.

The second one happens when trying to transform one object to another.
Since we cannot validate the `when` directives at configuration time,
it is possible that two are equivalent (for one or more cases). Up until
now, the first one that popped up would've been chosen. Instead, we'll
now make sure that this case blows up since this would lead to problems
of chosing the wrong mapping "randomly".