Provide a bit more context on transformation error
When an error/exception occurrs while transforming an element, there's
little information present apart from the error message itself. This
initially only reports on what the aprent transformation was, but we
unfortunately do no have any information about which properties are
being transformed as that is all handled inside a lambda.