Allow assigning role explicity in assembly connector
This extends the grammar to allow for specifying the required interface
an assembly connector would like to use. When a component requires an
interface multiple times, it was not possible to define which of them
should be used.

Additionally, this tries to automatically assign the correct target role
for the interface, if it is at all possible. Such as, if context A
provides interface I1 and is being connected to context B which requires
I1. In that case the connection will be established automatically.
However, if B now required I1 two times, we would now error out and
request the user to specify which role context A should be assigned to.
Similarly, if the provided roles of the providing context don't overlap
with the required roles in the requiring context, we also now show an
error to the user.