Fix inconsistent ID resolution in generated files
Currently, when we generate the output files, we both add them to the
resource set as well as save them in one go. This is problematic because
if a referenced element is not yet present in the resource set, it
cannot be resolved and results in a missing reference in the file.

Instead, all elements need to be added to the resource set first and
only then all files may be saved. As a result, all elements can now be
found.