Fix possibly using 'null' in the generated filename
When generating the name of the file that is the result of the
transformation, it was possible that it would contain 'null'. Reason
being that we allow elements to not have a name but we use that name
when generating the corresponding file. In order to avoid this, we fall
back to the type of the element that is described.