Use all input files when generating
Normally, the generator only gets passed in the file that needs to be
re-generated. This unfortunately does not work well with imports. With
imports, we'd need to add the target resources to our resource set so
when we save the resource that we just transformed, it knows where to
find the referenced elements. For now, this is achieved by just
generating all of the input files together forming a single resource set
as the ouput. While this has obviously downsides (e.g. if a file is not
referenced by anything else, we'd still generate all other files), it
lays the foundation for further improvements.