Inject registry via guice and allow for different providers
This also makes sure to inject the registry via the DI mechanism
provided by guice, as has been done for other parts in the past.

Additionally, the mechanism whichs configures the transformation
registry has been adjusted to also work without eclipse. This has
previously been handled exclusively by an extension point mechanism.
That is however not easy to set up in tests and we also want to allow
for different mechanisms in other cases as well. Thus, this is now a
more general provider setup which can provide the configurers as it
likes with the default being the extension point setup we had before.