Fix stack overflow because of infinite recursion
When trying to transform a datatype that looks like so:

```
datatype Composed {
    member1 datatype InnerCollection[]<Composed>
}
```

it would result in a stackoverflow. This was because when transforming
`Composed` it would go through all its members and transform them too.
Since we don't differentiate between a reference and a definition, the
`Composed` type specified in the `InnerCollection` will also be tried to
be transformed. Since that element has not yet finished transforming, we
continue in this endless loop.

We avoid this by saving the instance before we transform any of the
child properties. So once we do transform the properties and encouter
the parent element, that element is already remembered and will be
returned directly, instead of attempting to transfrom it.