Try and get type of the predefined mapped value
We allow for specifying already existing mappings inside a "mappings"
block. This allows us to specify already existing Palladio elements that
are special, such as the cpu resource interface. To do that, we provide
the element that should be replaced with them and then just create an
entry in the internal mapping cache. Once the element needs to be
transformed, it would check that cache and notice an element already
exists and use that instead of trying to create a completely new one.

However, we didn't specify a type for these previously. For many of the
possible mappings, this is fine and we'd still find the given element.
If a specific target is defined this no longer works. Now, one could
change the lookup to also include entries without a target, but at that
point we'd just be looking through the whole map every time. It just
makes more sense to instead specify a type when adding these entries to
the cache in the first place as it requires no changes in the way these
are looked up.