Do not include external resource interfaces
This is unfortunately a symptom of a bigger problem, but there doesn't
seem to be a better answer currently.

When we specify an already existing transformation (using the "mappings"
directive) we skip the recrusive transformation part and just return the
object. The function to set these values on the object still gets
called, as it may just be a one-directional reference. But in the case
of a parent-child relationship, like it is with resource interfaces in a
resource type repository, this is not what we want as the relationship
is already established. Running this function again causes the
relationship to be broken and we end up with an element that doesn't
belong.

By specifically ignoring elements which already have a parent, in this
case a resource type repository, assigned, we avoid this problem. But,
this would need to be done in every possible instance and is therefor in
the need of a better solution. For now however, this is a "fix".