Avoid adding unwanted, already mapped elements
As we've fixed in the previous commit, this avoids adding elements that
already exist under a different resource set that have been mapped to
elements in this resource set. Otherwise, if we add them, they'd get
ripped out and then end up being lost. We only want to have those when
assigning them to specific properties not as elements of a container.

The problem we didn't address in the last commit was that manually
filtering the elements in each configuration was error prone and
cumbersome and thus we only implemented it for that one case that we
currently had problems with. We acknowledged it would probably cause
issues the next time someone tried to map something however. This is now
easier than anticipated: we didn't need to set the backwards reference
in those bidirectional references. This is done internally automatically
and we thus only need one call. as a result, we can remove quite a bit
of code everwhere but it also allows us to create a simple extension
method to deal with the problem. It is now elegant and straight forward.

We do not need to apply a similar approach to single references. Simply
put, these do not have this issue.