Added writeup about the addon
This writeup contains information about how the addon is structured,
what components it contains and how it in general works. It does not
touch the grammar itself very much as it is mostly a straight forward
mapping of grammar element to palladio attribute, but instead explains
the transformation mechanism and how rules are implemented.